+++ tests/test_pyseto.py
@@ -121,6 +121,59 @@ class TestPyseto:
         assert msg in str(err.value)
 
     @pytest.mark.parametrize(
+        "serializer, msg",
+        [
+            (
+                None,
+                "serializer should be specified for the footer object.",
+            ),
+            (
+                {},
+                "serializer should be specified for the footer object.",
+            ),
+            (
+                [],
+                "serializer should be specified for the footer object.",
+            ),
+            (
+                "",
+                "serializer should be specified for the footer object.",
+            ),
+            (
+                b"",
+                "serializer should be specified for the footer object.",
+            ),
+            (
+                {"key": "value"},
+                "serializer should have dumps().",
+            ),
+            (
+                InvalidSerializer(),
+                "serializer should have dumps().",
+            ),
+            (
+                InvalidSerializer2(),
+                "Failed to serialize the footer.",
+            ),
+        ],
+    )
+    def test_encode_object_footer_with_invalid_serializer(self, serializer, msg):
+        private_key_pem = b"-----BEGIN PRIVATE KEY-----\nMC4CAQAwBQYDK2VwBCIEILTL+0PfTOIQcn2VPkpxMwf6Gbt9n4UEFDjZ4RuUKjd0\n-----END PRIVATE KEY-----"
+
+        private_key = Key.new(version=4, purpose="public", key=private_key_pem)
+        with pytest.raises(ValueError) as err:
+            pyseto.encode(
+                private_key,
+                b"Hello world!",
+                footer={
+                    "kid": "xxxxxx",
+                },
+                serializer=serializer,
+            )
+            pytest.fail("pyseto.encode() should fail.")
+        assert msg in str(err.value)
+
+    @pytest.mark.parametrize(
         "version, key, msg",
         [
             (1, load_key("keys/public_key_rsa.pem"), "Invalid payload."),
@@ -184,6 +237,20 @@ class TestPyseto:
             pytest.fail("pyseto.decode() should fail.")
         assert msg in str(err.value)
 
+    def test_decode_bytes_footer_with_deserializer(self):
+        private_key_pem = b"-----BEGIN PRIVATE KEY-----\nMC4CAQAwBQYDK2VwBCIEILTL+0PfTOIQcn2VPkpxMwf6Gbt9n4UEFDjZ4RuUKjd0\n-----END PRIVATE KEY-----"
+        public_key_pem = b"-----BEGIN PUBLIC KEY-----\nMCowBQYDK2VwAyEAHrnbu7wEfAP9cGBOAHHwmH4Wsot1ciXBHwBBXQ4gsaI=\n-----END PUBLIC KEY-----"
+        private_key = Key.new(version=4, purpose="public", key=private_key_pem)
+        token = pyseto.encode(
+            private_key,
+            {"data": "this is a signed message", "exp": "2022-01-01T00:00:00+00:00"},
+            footer=b"This is a footer.",
+        )
+        public_key = Key.new(version=4, purpose="public", key=public_key_pem)
+        decoded = pyseto.decode(public_key, token, deserializer=json)
+        assert isinstance(decoded.footer, bytes)
+        assert decoded.footer == b"This is a footer."
+
     def test_decode_object_payload_with_invalid_exp(self):
 
         private_key_pem = b"-----BEGIN PRIVATE KEY-----\nMC4CAQAwBQYDK2VwBCIEILTL+0PfTOIQcn2VPkpxMwf6Gbt9n4UEFDjZ4RuUKjd0\n-----END PRIVATE KEY-----"
@@ -192,7 +259,7 @@ class TestPyseto:
         private_key = Key.new(version=4, purpose="public", key=private_key_pem)
         token = pyseto.encode(
             private_key,
-            {"data": "this is a signed message", "exp": "xxxxx"},
+            {"data": "this is a signed message"},
         )
         public_key = Key.new(version=4, purpose="public", key=public_key_pem)
         with pytest.raises(VerifyError) as err: