+++ pyseto/__init__.py
@@ -10,7 +10,7 @@ from .key import Key
 from .paseto import Paseto
 from .pyseto import decode, encode
 
-__version__ = "1.2.0"
+__version__ = "1.3.0"
 __title__ = "PySETO"
 __description__ = "A Python implementation of PASETO"
 __url__ = "https://pyseto.readthedocs.io"