+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changes
 Unreleased
 ----------
 
+- Add support for exp claim. `#71 <https://github.com/dajiaji/pyseto/pull/71>`__
+
 Version 1.2.0
 -------------