+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changes
 Unreleased
 ----------
 
+- Add leeway for exp validation. `#73 <https://github.com/dajiaji/pyseto/pull/73>`__
 - Add Paseto class. `#72 <https://github.com/dajiaji/pyseto/pull/72>`__
 - Add support for exp claim. `#71 <https://github.com/dajiaji/pyseto/pull/71>`__